add ignore product image uploaded in .gitignore, add group for route kasir, build buy product system, cart system, make all model not use autoincrement, add array product_price_formatted and rename number_product to product_sales in array in remapDataProducts method, add generate_uuid helper, remove POSWModel.php, replace code 'echo json_encode...' with 'return json_encode...' in all method that before use it, change model_price to price_model in constructor method controller product, add is_unique rule and change if condition from '=== true' to '> 0' in condition 'if insert product and insert product price success' in saveProductToDB and updateProductInDB method, refactory variabel name and id for tag html, remove use try block in updateProductInDB method, add condition for check is file exists if product update success, change update product message if not success, replace condition ' === true' to '> 0' in updateProductInDB method, replace condition for check remove product success from '=== true' to '> 0' in removeProductInDB method, remove default value for primary key in migrations, make BaseModel, change visibility property table to protected in all model, change return type hinting in removeCategoryProduct, removeProductPrice, removeUser and removeProduct method to int, make transactionDetailModel and transactionModel, change statis user sign in name to dynamic from session, refactory condition for check if exists bestseller product in cashier.php, make cashier_js.php, transalate error page for production to indonesian language, change condition for check if product exists in search product in product.php, refactory condition for check product magnitude old > product price db in update_product.php and function number_formatter_to_currency in posw.js

+++ app/Controllers/Product.php
@@ -2,7 +2,6 @@
 
 use CodeIgniter\Controller;
 use App\Models\CategoryProductModel;
-use App\Models\POSWModel;
 use App\Models\ProductModel;
 use App\Models\ProductPriceModel;
 use App\Libraries\ValidationMessage;
@@ -10,14 +9,14 @@ use CodeIgniter\HTTP\Files\UploadedFile;
 
 class Product extends Controller
 {
-    protected $helpers = ['form', 'active_menu', 'check_password_sign_in_user'];
+    protected $helpers = ['form', 'active_menu', 'check_password_sign_in_user', 'generate_uuid'];
     private const PRODUCT_LIMIT = 50;
 
     public function __construct()
     {
         $this->session = \Config\Services::session();
         $this->model = new ProductModel();
-        $this->model_price = new ProductPriceModel();
+        $this->price_model = new ProductPriceModel();
     }
 
     public function index()
@@ -108,6 +107,7 @@ class Product extends Controller
         $count_product_magnitude = count($product_magnitudes);
         for ($i = 0; $i < $count_product_magnitude; $i++) {
             $data_insert[] = [
+                'harga_produk_id' => generate_uuid(),
                 'produk_id' => $product_id,
                 'besaran_produk' => filter_var($product_magnitudes[$i], FILTER_SANITIZE_STRING),
                 'harga_produk' => filter_var($product_prices[$i], FILTER_SANITIZE_STRING)
@@ -128,8 +128,8 @@ class Product extends Controller
             ],
             'product_name' => [
                 'label' => 'Nama produk',
-                'rules' => 'required|max_length[50]',
-                'errors' => ValidationMessage::generateIndonesianErrorMessage('required','max_length')
+                'rules' => 'required|max_length[50]|is_unique[produk.nama_produk]',
+                'errors' => ValidationMessage::generateIndonesianErrorMessage('required','max_length','is_unique')
             ],
             'product_status' => [
                 'label' => 'Status Produk',
@@ -181,8 +181,8 @@ class Product extends Controller
 
         $this->model->db->transStart();
         // insert product
-        $posw_model_product = new POSWModel($this->model->db, $this->model->table);
-        $insert_product = $posw_model_product->insertReturning([
+        $insert_product = $this->model->insertReturning([
+            'produk_id' => generate_uuid(),
             'kategori_produk_id' => $this->request->getPost('category_product', FILTER_SANITIZE_STRING),
             'nama_produk' => $this->request->getPost('product_name', FILTER_SANITIZE_STRING),
             'foto_produk' => $product_photo_name,
@@ -191,15 +191,14 @@ class Product extends Controller
         ], 'produk_id');
 
         // insert product price
-        $produk_id = $posw_model_product->getInsertReturning()??'';
-        $data_product_prices = $this->generateDataInsertBatchProductPrice($produk_id, $product_magnitudes, $product_prices);
-        $posw_model_product_price = new POSWModel($this->model->db, 'harga_produk');
-        $insert_product_price = $posw_model_product_price->insertBatch($data_product_prices);
+        $produk_id = $this->model->getInsertReturned()??'';
+        $data_product_price = $this->generateDataInsertBatchProductPrice($produk_id, $product_magnitudes, $product_prices);
+        $insert_product_price = $this->price_model->insertBatch($data_product_price);
 
         $this->model->db->transComplete();
 
         // if insert product and insert product price success
-        if ($insert_product === true && $insert_product_price === true) {
+        if ($insert_product === true && $insert_product_price > 0) {
             // move product photo
             $product_photo_file->move('dist/images/product_photo', $product_photo_name);
 
@@ -219,11 +218,10 @@ class Product extends Controller
     public function showProductDetail()
     {
         $product_id = $this->request->getPost('product_id', FILTER_SANITIZE_STRING);
-        $product_prices = $this->model_price->getProductPrices($product_id, 'harga_produk, besaran_produk');
+        $product_prices = $this->price_model->getProductPrices($product_id, 'harga_produk, besaran_produk');
         $product_photo = $this->model->findProduct($product_id, 'foto_produk')['foto_produk']??null;
 
-        echo json_encode(['product_prices'=>$product_prices, 'product_photo'=>$product_photo, 'csrf_value'=>csrf_hash()]);
-        return true;
+        return json_encode(['product_prices'=>$product_prices, 'product_photo'=>$product_photo, 'csrf_value'=>csrf_hash()]);
     }
 
     public function showProductSearches()
@@ -241,13 +239,12 @@ class Product extends Controller
         // get product search total
         $product_search_total = $this->model->countAllProductSearch($keyword);
 
-        echo json_encode([
+        return json_encode([
             'products_db' => $products_db,
             'product_search_total' => $product_search_total,
             'product_limit' => static::PRODUCT_LIMIT,
             'csrf_value' => csrf_hash()
         ]);
-        return true;
     }
 
     public function updateProduct(string $product_id)
@@ -260,7 +257,7 @@ class Product extends Controller
         $data['page'] = 'perbaharui_produk';
         $data['product_id'] = $product_id;
         $data['product_db'] = $this->model->findProduct($product_id, 'kategori_produk_id,nama_produk,status_produk,foto_produk');
-        $data['product_prices_db'] = $this->model_price->getProductPrices($product_id, 'harga_produk_id, besaran_produk, harga_produk');
+        $data['product_prices_db'] = $this->price_model->getProductPrices($product_id, 'harga_produk_id, besaran_produk, harga_produk');
         $data['category_products_db'] = $category_product_model->getCategoryProductsForFormSelect();
 
         return view('product/update_product', $data);
@@ -268,28 +265,28 @@ class Product extends Controller
 
     private function splitProductPriceCreateUpdate(array $product_price_ids, array $product_magnitudes, array $product_prices)
     {
-        $product_magnitudes_update = [];
-        $product_prices_update = [];
-        $product_magnitudes_insert = [];
-        $product_prices_insert = [];
+        $data_product_magnitude_update = [];
+        $data_product_price_update = [];
+        $data_product_magnitude_insert = [];
+        $data_product_price_insert = [];
 
         $count_product_magnitude = count($product_magnitudes);
         for ($i = 0; $i < $count_product_magnitude; $i++) {
             // if product_price_id exists
             if (isset($product_price_ids[$i])) {
-                $product_magnitudes_update[] = $product_magnitudes[$i];
-                $product_prices_update[] = $product_prices[$i];
+                $data_product_magnitude_update[] = $product_magnitudes[$i];
+                $data_product_price_update[] = $product_prices[$i];
             } else {
-                $product_magnitudes_insert[] = $product_magnitudes[$i];
-                $product_prices_insert[] = $product_prices[$i];
+                $data_product_magnitude_insert[] = $product_magnitudes[$i];
+                $data_product_price_insert[] = $product_prices[$i];
             }
         }
 
         return [
-            'product_magnitudes_update' => $product_magnitudes_update,
-            'product_prices_update' => $product_prices_update,
-            'product_magnitudes_insert' => $product_magnitudes_insert,
-            'product_prices_insert' => $product_prices_insert
+            'data_product_magnitude_update' => $data_product_magnitude_update,
+            'data_product_price_update' => $data_product_price_update,
+            'data_product_magnitude_insert' => $data_product_magnitude_insert,
+            'data_product_price_insert' => $data_product_price_insert
         ];
     }
 
@@ -365,10 +362,10 @@ class Product extends Controller
             return redirect()->back()->withInput();
         }
 
-        $data_update_product = [];
+        $data_product_update = [];
         $product_id = $this->request->getPost('product_id',FILTER_SANITIZE_STRING);
 
-        $data_update_product = array_merge($data_update_product, [
+        $data_product_update = array_merge($data_product_update, [
             'kategori_produk_id' => $this->request->getPost('category_product', FILTER_SANITIZE_STRING),
             'nama_produk' => $this->request->getPost('product_name', FILTER_SANITIZE_STRING),
             'status_produk' => $this->request->getPost('product_status', FILTER_SANITIZE_STRING),
@@ -379,7 +376,7 @@ class Product extends Controller
         if ($product_photo_file->getError() !== 4) {
             $product_photo_name = $product_photo_file->getRandomName();
             // genearate new random name for product photo
-            $data_update_product = array_merge($data_update_product, ['foto_produk' => $product_photo_name]);
+            $data_product_update = array_merge($data_product_update, ['foto_produk' => $product_photo_name]);
         }
 
         $product_price_ids = $this->request->getPost('product_price_ids');
@@ -390,51 +387,42 @@ class Product extends Controller
             $this->request->getPost('product_prices')
         );
 
-        $product_magnitudes_update = $split_product_price['product_magnitudes_update'];
-        $product_prices_update = $split_product_price['product_prices_update'];
-        $product_magnitudes_insert = $split_product_price['product_magnitudes_insert'];
-        $product_prices_insert = $split_product_price['product_prices_insert'];
+        $data_product_magnitude_update = $split_product_price['data_product_magnitude_update'];
+        $data_product_price_update = $split_product_price['data_product_price_update'];
+        $data_product_magnitude_insert = $split_product_price['data_product_magnitude_insert'];
+        $data_product_price_insert = $split_product_price['data_product_price_insert'];
 
         $data_product_price_update = $this->generateDataUpdateBatchProductPrice(
             $product_price_ids,
-            $product_magnitudes_update,
-            $product_prices_update
+            $data_product_magnitude_update,
+            $data_product_price_update
         );
 
         $data_product_price_insert = $this->generateDataInsertBatchProductPrice(
             $product_id,
-            $product_magnitudes_insert,
-            $product_prices_insert
+            $data_product_magnitude_insert,
+            $data_product_price_insert
         );
 
-        try {
-            $this->model->transBegin();
-            // update product
-            $update_product = $this->model->update($product_id, $data_update_product);
+        $this->model->transStart();
 
-            // update product price
-            $update_product_price = $this->model_price->updateBatch($data_product_price_update, 'harga_produk_id');
-
-            // insert product price
-            if (count($data_product_price_insert) > 0) {
-                $posw_model_product_price = new POSWModel($this->model->db, 'harga_produk');
-                $insert_product_price = $posw_model_product_price->insertBatch($data_product_price_insert);
-            }
-
-            // commit transaction
-            $this->model->transCommit();
-
-        } catch (\ErrorException $e) {
-            // rollback transaction
-            $this->model->transRollBack();
+        // update product
+        $update_product = $this->model->update($product_id, $data_product_update);
+        // update product price
+        $update_product_price = $this->price_model->updateBatch($data_product_price_update, 'harga_produk_id');
+        // insert product price
+        if (count($data_product_price_insert) > 0) {
+            $insert_product_price = $this->price_model->insertBatch($data_product_price_insert);
         }
+
+        $this->model->transComplete();
 
         /* if update product and update product price success and
          *  data insert product price not found or insert product price success */
         if (
             $update_product === true &&
             $update_product_price === count($data_product_price_update) &&
-            (count($data_product_price_insert) === 0 || $insert_product_price === true)
+            (count($data_product_price_insert) === 0 || $insert_product_price > 0)
         ) {
             // if exists product photo
             if ($product_photo_file->getError() === 0) {
@@ -442,7 +430,9 @@ class Product extends Controller
                 $product_photo_file->move('dist/images/product_photo', $product_photo_name);
                 // remove old product photo
                 $old_product_photo = $this->request->getPost('old_product_photo');
-                unlink('dist/images/product_photo/'.$old_product_photo);
+                if (file_exists('dist/images/product_photo/'.$old_product_photo)) {
+                    unlink('dist/images/product_photo/'.$old_product_photo);
+                }
             }
 
             // make success message
@@ -461,7 +451,7 @@ class Product extends Controller
             'form_errors',
             '<div class="alert alert--warning mb-3"><span class="alert__icon"></span><p>',
             '</p><a class="alert__close" onclick="close_alert(event)" href="#"></a></div>',
-            ['update_product' => '<strong>Peringatan</strong>, Produk gagal dibuat']
+            ['update_product' => '<strong>Peringatan</strong>, Produk gagal diperbaharui']
         );
         return redirect()->back();
     }
@@ -469,14 +459,12 @@ class Product extends Controller
     public function removeProductPriceInDB()
     {
         $product_price_id = $this->request->getPost('product_price_id', FILTER_SANITIZE_STRING);
-        if ($this->model_price->removeProductPrice($product_price_id) === true) {
-            echo json_encode(['success'=>true, 'csrf_value'=>csrf_hash()]);
-            return true;
+        if ($this->price_model->removeProductPrice($product_price_id) > 0) {
+            return json_encode(['success'=>true, 'csrf_value'=>csrf_hash()]);
         }
 
         $error_message = 'Gagal menghapus harga produk, cek apakah masih ada transaksi yang terhubung!';
-        echo json_encode(['success'=>false, 'error_message'=>$error_message, 'csrf_value'=>csrf_hash()]);
-        return false;
+        return json_encode(['success'=>false, 'error_message'=>$error_message, 'csrf_value'=>csrf_hash()]);
     }
 
     public function removeProductInDB()
@@ -484,10 +472,12 @@ class Product extends Controller
         $product_ids = explode(',',$this->request->getPost('product_ids', FILTER_SANITIZE_STRING));
         $photo_products = $this->model->findProducts($product_ids, 'foto_produk');
         // remove product
-        if ($this->model->removeProduct($product_ids) === true) {
+        if ($this->model->removeProduct($product_ids) > 0) {
             // remove photo product
             foreach($photo_products as $p) {
-                unlink('dist/images/product_photo/'.$p['foto_produk']);
+                if (file_exists('dist/images/product_photo/'.$p['foto_produk'])) {
+                    unlink('dist/images/product_photo/'.$p['foto_produk']);
+                }
             }
 
             $count_product_ids = count($product_ids);
@@ -515,18 +505,16 @@ class Product extends Controller
                 $longer_products[$i]['waktu_buat_indo'] = $date_time->convertTimstampToIndonesianDateTime($longer_products[$i]['waktu_buat']);
             }
 
-            echo json_encode([
+            return json_encode([
                 'success' => true,
                 'longer_products' => $longer_products,
                 'product_total' => $product_total,
                 'product_limit' => static::PRODUCT_LIMIT,
                 'csrf_value' => csrf_hash()
             ]);
-            return true;
         }
 
-        $error_message = 'Gagal menghapus product, cek apakah masih ada transaksi yang terhubung!';
-        echo json_encode(['success'=>false, 'error_message'=>$error_message, 'csrf_value'=>csrf_hash()]);
-        return false;
+        $error_message = 'Gagal menghapus produk, cek apakah masih ada transaksi yang terhubung!';
+        return json_encode(['success'=>false, 'error_message'=>$error_message, 'csrf_value'=>csrf_hash()]);
     }
 }
